csis-192: For FS vols, move globalstorage creation.
There is a globalmount file that contains persistent meta data about the
volume needed to remove it later.   That was getting created with a
missing mpath field if the volume never got mounted.

Create this file directly before attempting the first mount so that if
the mount fails, it will still exist.

DCO 1.1 Signed-off-by: David Ohlemacher <dohlemacher@infinidat.com>