csic-320: Fix iSCSI block DetachDisk.
Do not try to check the mntPath or targetPath.
Instead, just call Unmount and handle errors if the occur
(aka ask forgiveness rather than permission).

This bug is caused by the confusing use of "/host" in some, but not all
paths and chown in the Dockerfile.

DCO 1.1 Signed-off-by: David Ohlemacher <dohlemacher@infinidat.com>